Crate ic_interfaces

The public interfaces of the various components of the replica.

Having the public interfaces defined separately from their components helps reduce unnecessary dependencies between them.

Modules

artifact_manager
The artifact manager/client public interface.

artifact_pool
The artifact pool public interface.

certification
The certification public interface.

certified_stream_store
The certified stream store public interface.

consensus
The consensus public interface.

consensus_pool
The consensus pool public interface.

crypto
The crypto public interface.

dkg
The DKG public interface.

execution_environment
The execution environment public interface.

gossip_pool
The gossip pool public interface.

ingress_manager
The ingress manager public interface.

ingress_pool
The ingress pool public interface.

messages
Messages used in various components.

messaging
Message Routing public interfaces.

p2p
The P2P public interface.

registry
The registry public interface.

replica_config
The replica config store.

state_manager
The state manager public interface.

time_source
The time source public interface.

transport
Transport layer public interface.

validation
Validation related types.