Trait ic_interfaces::artifact_manager::ArtifactClient

pub trait ArtifactClient<Artifact: ArtifactKind>: Send + Sync {
    fn check_artifact_acceptance(
        &self, 
        msg: Artifact::SerializeAs, 
        peer_id: &NodeId
    ) -> Result<ArtifactAcceptance<Artifact::Message>, ArtifactPoolError>;

    fn has_artifact(&self, msg_id: &Artifact::Id) -> bool;

    fn get_validated_by_identifier(
        &self, 
        msg_id: &Artifact::Id
    ) -> Option<Artifact::Message>;

    fn get_priority_function(
        &self
    ) -> Option<PriorityFn<Artifact::Id, Artifact::Attribute>>;

    fn get_chunk_tracker(
        &self, 
        artifact_id: &Artifact::Id
    ) -> Box<dyn Chunkable + Send + Sync>;

    fn get_filter(&self) -> Artifact::Filter
    where
        Artifact::Filter: Default,
    { ... }

    fn get_all_validated_by_filter(
        &self, 
        _filter: &Artifact::Filter
    ) -> Vec<Advert<Artifact>> { ... }

    fn get_remaining_quota(&self, _peer_id: NodeId) -> usize { ... }
}

An abstraction of artifact processing for a sub-type of the overall 'Artifact' type.

Required methods

fn check_artifact_acceptance(
&self,
msg: Artifact::SerializeAs,
peer_id: &NodeId
) -> Result<ArtifactAcceptance<Artifact::Message>, ArtifactPoolError>

When a new artifact is received, check_artifact_acceptance function is called to perform basic pre-processing and check if the artifact matches the advert used to request it if one is passed along. Note that this function should not modify the artifact pool.

If it passes the pre-processing, the same artifact should be returned, which will then be passed on to the corresponding 'ArtifactProcessor' component afterwards. Otherwise it will be rejected.

The default implementation is to accept unconditionally.

fn has_artifact(&self, msg_id: &Artifact::Id) -> bool

Checks if the node already has the artifact in the pool by its identifier.

fn get_validated_by_identifier(
&self,
msg_id: &Artifact::Id
) -> Option<Artifact::Message>

Gets a validated artifact by its identifier. Return None if no valid artifact is found for the given identifier.

fn get_priority_function(
&self
) -> Option<PriorityFn<Artifact::Id, Artifact::Attribute>>

Return the priority function used by this client.

fn get_chunk_tracker(
&self,
artifact_id: &Artifact::Id
) -> Box<dyn Chunkable + Send + Sync>

Get Chunk tracker for an advert. Download/Chunk trackers for Semi-structured/multi-chunk artifacts need to be operated by pool clients. Clients own the tracking logic, this callback is for them to setup chunk iterator context etc. For example This call may be used by an artifact with on-disk chunks to setup the directory and iterator logic before gossip starts calling into the iterator.

Loading content...
Provided methods

fn get_filter(&self) -> Artifact::Filter where
Artifact::Filter: Default,

Gets the filter that needs to be sent with re-transmission request to other peers. This filter contains the information to indicate to other peers what this peer already has, and what relevant adverts are still needed.

Assuming that Node A is a node trying to resume with the help of Node B. get_filter will be used by Node A to figure out its current filter and will be sent to Node B as a part of the re-transmission request. Node B will then use get_all_validated_by_filter with that filter to derive only the relevant adverts to be sent back to Node A.

In the first version to be implemented, this filter is the last finalized height of Consensus. For all pool handlers this can be used to derive a suitable threshold value. For Consensus it will be height, for Ingress messages the time stamp of the finalized block at this height, for DKG the DKG instance relevant at this height, for Certification and State Synchronization the latest available executed state referred to in the block at this height.

Example If Consensus pool has delivered batches up to height 10, the filter will be 'height = 10' since this node only needs consensus artifacts with height > 10.

fn get_all_validated_by_filter(
&self,
_filter: &Artifact::Filter
) -> Vec<Advert<Artifact>>

Get adverts of all validated artifacts by the filter. This filter is used to derive only the relevant adverts.

Assuming that Node A is a node trying to resume with the help of Node B. get_filter will be used by Node A to figure out its current filter and will be sent to Node B as a part of the re-transmission request. Node B will then use get_all_validated_by_filter with that filter to derive only the relevant adverts to be sent back to Node A.

Example If the filter contains height = 10; adverts for all the validated artifacts with height > 10 will be returned by this function.

fn get_remaining_quota(&self, _peer_id: NodeId) -> usize

Return the remaining quota that this peer is allowed to consume. Return None if the quota is not specified.

Loading content...