Trait ic_interfaces::artifact_manager::ArtifactManager

pub trait ArtifactManager: Send + Sync {
    fn on_artifact(
        &self, 
        msg: Artifact, 
        advert: GossipAdvert, 
        peer_id: &NodeId
    ) -> Result<(), OnArtifactError<Artifact>>;

    fn has_artifact(&self, message_id: &ArtifactId) -> bool;

    fn get_validated_by_identifier(
        &self, 
        message_id: &ArtifactId
    ) -> Option<Box<dyn ChunkableArtifact>>;

    fn get_filter(&self) -> ArtifactFilter;

    fn get_all_validated_by_filter(
        &self, 
        filter: &ArtifactFilter
    ) -> Vec<GossipAdvert>;

    fn get_remaining_quota(
        &self, 
        tag: ArtifactTag, 
        peer_id: NodeId
    ) -> Option<usize>;

    fn get_priority_function(
        &self, 
        tag: ArtifactTag
    ) -> Option<ArtifactPriorityFn>;

    fn get_chunk_tracker(
        &self, 
        artifact_id: &ArtifactId
    ) -> Option<Box<dyn Chunkable + Send + Sync>>;
}

The Artifact Manager stores artifacts to be used by this and other nodes in the same subnet in the artifact pool.

Required methods

fn on_artifact(
&self,
msg: Artifact,
advert: GossipAdvert,
peer_id: &NodeId
) -> Result<(), OnArtifactError<Artifact>>

When a new artifact is received by Gossip, it is forwarded to the ArtifactManager together with its advert via the on_artifact call. This then forwards them to be processed by the corresponding ArtifactClient/ArtifactProcessor based on the artifact type. Returns OnArtifactError if no clients were able to process it or an error has occurred.

See ArtifactClient::on_artifact for more details.

fn has_artifact(&self, message_id: &ArtifactId) -> bool

Checks if any of the ArtifactClient already has the artifact in the pool by the given identifier.

fn get_validated_by_identifier(
&self,
message_id: &ArtifactId
) -> Option<Box<dyn ChunkableArtifact>>

Return a validated artifact by its identifier, or None if not found.

fn get_filter(&self) -> ArtifactFilter

Gets the filter that needs to be sent with re-transmission request to other peers. This filter should be a collection of all filters returned by the ArtifactClients.

See ArtifactClient::get_filter for more details.

fn get_all_validated_by_filter(
&self,
filter: &ArtifactFilter
) -> Vec<GossipAdvert>

Get adverts of all validated artifacts by the filter from all clients.

See ArtifactClient::get_all_validated_by_filter for more details.

fn get_remaining_quota(
&self,
tag: ArtifactTag,
peer_id: NodeId
) -> Option<usize>

Gets the remaining quota the given peer is allowed to consume for a specific client that is identified by the given artifact tag.

See ArtifactClient::get_remaining_quota for more details.

fn get_priority_function(&self, tag: ArtifactTag) -> Option<ArtifactPriorityFn>

Return the priority function for a specific client that is identified by the given artifact tag.

See ArtifactClient::get_priority_function for more details.

fn get_chunk_tracker(
&self,
artifact_id: &ArtifactId
) -> Option<Box<dyn Chunkable + Send + Sync>>

Get Chunk tracker for an advert.

See ArtifactClient::get_chunk_tracker for more details

Loading content...