Module ic_interfaces::artifact_pool

The artifact pool public interface.

Structs

ReplicaVersionMismatch
Describe expected version and artifact version when there is a mismatch.

UnvalidatedArtifact
Unvalidated artifact

ValidatedArtifact
Validated artifact

Enums

ArtifactPoolError

Traits

HasTimestamp
A trait to get timestamp.

IntoInner
A trait similar to Into, but without its restrictions.