Trait ic_interfaces::certification::CertificationPool

pub trait CertificationPool {
    fn certification_at_height(&self, height: Height) -> Option<Certification>;

    fn shares_at_height(
        &self, 
        height: Height
    ) -> Box<dyn Iterator<Item = CertificationShare>>;

    fn validated_shares(&self) -> Box<dyn Iterator<Item = CertificationShare>>;

    fn unvalidated_certifications_at_height(
        &self, 
        height: Height
    ) -> Box<dyn Iterator<Item = &Certification>>;

    fn unvalidated_shares_at_height(
        &self, 
        height: Height
    ) -> Box<dyn Iterator<Item = &CertificationShare>>;

    fn all_heights_with_artifacts(&self) -> Vec<Height>;

    fn certified_heights(&self) -> HashSet<Height>;
}

Trait containing only immutable functions wrt. Certification Pool

Required methods

fn certification_at_height(&self, height: Height) -> Option<Certification>

Returns the certification for the given height, if available.

fn shares_at_height(
&self,
height: Height
) -> Box<dyn Iterator<Item = CertificationShare>>

Returns an iterator over all shares for the given height.

fn validated_shares(&self) -> Box<dyn Iterator<Item = CertificationShare>>

Returns all validated certification shares.

fn unvalidated_certifications_at_height(
&self,
height: Height
) -> Box<dyn Iterator<Item = &Certification>>

Returns an iterator of all unvalidated full certification for the given height.

fn unvalidated_shares_at_height(
&self,
height: Height
) -> Box<dyn Iterator<Item = &CertificationShare>>

Returns an iterator of all unvalidated shares for the given height.

fn all_heights_with_artifacts(&self) -> Vec<Height>

Returns all heights (sorted, ascending), where at least a single artifact is present. Intended to be used by the purger.

fn certified_heights(&self) -> HashSet<Height>

Returns all heights which have a full validated certification.

Loading content...