Trait ic_interfaces::certification::CertifierGossip

pub trait CertifierGossip: Send + Sync {
    fn get_priority_function(
        &self, 
        consensus_cache: &dyn ConsensusPoolCache, 
        certification_pool: &dyn CertificationPool
    ) -> PriorityFn<CertificationMessageId, CertificationMessageAttribute>;

    fn get_filter(&self) -> CertificationMessageFilter;
}

Trait containing methods related to gossiping.

Required methods

fn get_priority_function(
&self,
consensus_cache: &dyn ConsensusPoolCache,
certification_pool: &dyn CertificationPool
) -> PriorityFn<CertificationMessageId, CertificationMessageAttribute>

Return the priority function for the Gossip protocol to optimize the artifact exchange.

fn get_filter(&self) -> CertificationMessageFilter

Return a filter that represents what artifacts are needed.

Loading content...