Trait ic_interfaces::certification::MutableCertificationPool

pub trait MutableCertificationPool: CertificationPool {
    fn insert(&mut self, msg: CertificationMessage);

    fn apply_changes(&mut self, change_set: ChangeSet);
}

Trait containing only mutable functions wrt. Certification Pool

Required methods

fn insert(&mut self, msg: CertificationMessage)

Inserts a certification message into the unvalidated part of the pool.

fn apply_changes(&mut self, change_set: ChangeSet)

Applies a set of change actions to the pool.

Loading content...