Trait ic_interfaces::certification::Verifier

pub trait Verifier: Send + Sync {
    fn validate(
        &self, 
        subnet_id: SubnetId, 
        certification: &Certification, 
        registry_version: RegistryVersion
    ) -> ValidationResult<VerifierError>;
}

Verifier is used to verify state hash certifications. It will be injected into XNet and StateSync components so that they can ensure the authenticity of parts of the replicated state sent over from other sub-networks.

Required methods

fn validate(
&self,
subnet_id: SubnetId,
certification: &Certification,
registry_version: RegistryVersion
) -> ValidationResult<VerifierError>

This method verifies whether the given certification contains a valid signature on the given hash on behalf of the subnet specified in subnet_id with respect to the registry_version.

Loading content...