Trait ic_interfaces::certified_stream_store::CertifiedStreamStore

pub trait CertifiedStreamStore: Send + Sync {
    fn encode_certified_stream_slice(
        &self, 
        remote_subnet: SubnetId, 
        witness_begin: Option<StreamIndex>, 
        msg_begin: Option<StreamIndex>, 
        msg_limit: Option<usize>, 
        byte_limit: Option<usize>
    ) -> Result<CertifiedStreamSlice, EncodeStreamError>;

    fn decode_certified_stream_slice(
        &self, 
        remote_subnet: SubnetId, 
        registry_version: RegistryVersion, 
        certified_slice: &CertifiedStreamSlice
    ) -> Result<StreamSlice, DecodeStreamError>;

    fn decode_valid_certified_stream_slice(
        &self, 
        certified_slice: &CertifiedStreamSlice
    ) -> Result<StreamSlice, DecodeStreamError>;

    fn subnets_with_certified_streams(&self) -> Vec<SubnetId>;
}

Responsible for fetching slices of certified streams and attaching certifications to them.

Such certifications allow subnets to verify that a slice fetched from an arbitrary node of a remote subnet is indeed genuine and agreed upon by a majority of the replicas constituting that subnet.

Required methods

fn encode_certified_stream_slice(
&self,
remote_subnet: SubnetId,
witness_begin: Option<StreamIndex>,
msg_begin: Option<StreamIndex>,
msg_limit: Option<usize>,
byte_limit: Option<usize>
) -> Result<CertifiedStreamSlice, EncodeStreamError>

Produces a certified slice of the stream for remote_subnet from the latest certified state, with a witness beginning at witness_begin; and messages beginning at msg_begin and containing at most msg_limit messages totaling at most byte_limit bytes.

Precondition: witness_begin.is_none() && msg_begin.is_none() || witness_begin.unwrap() <= msg_begin.unwrap().

fn decode_certified_stream_slice(
&self,
remote_subnet: SubnetId,
registry_version: RegistryVersion,
certified_slice: &CertifiedStreamSlice
) -> Result<StreamSlice, DecodeStreamError>

Decode the certified stream slice of the stream coming from the remote_subnet and validate the signature on it with respect to the given registry_version.

fn decode_valid_certified_stream_slice(
&self,
certified_slice: &CertifiedStreamSlice
) -> Result<StreamSlice, DecodeStreamError>

Decodes the certified stream slice without performing any validation. This method should only be used for decoding streams that have already been validated (e.g., payloads from previous blocks).

fn subnets_with_certified_streams(&self) -> Vec<SubnetId>

Returns the list of subnet ids for which we have outgoing certified streams.

Loading content...