Module ic_interfaces::consensus_pool

The consensus pool public interface.

Structs

ChainIterator
An iterator for block ancestors.

HeightRange

Enums

ChangeAction
Change actions applicable to the consensus pool.

OnlyError

Traits

ChangeSetOperation
A trait with common methods for change sets.

ConsensusPool
The consensus pool contains all the artifacts received by P2P and produced by the local node.

ConsensusPoolCache
Reader of consensus related states.

HeightIndexedPool
HeightIndexedPool provides a set of interfaces for the Consensus component to query artifacts. The same interface is applicable to both validated and unvalidated partitions of consensus artifacts in the overall ArtifactPool.

MutableConsensusPool
Mutation operations on top of ConsensusPool.

PoolSection
A Pool section is a part of the consensus pool which contains artifacts.

Type Definitions

ChangeSet

UnvalidatedConsensusArtifact
Unvalidated consensus artifact.

ValidatedConsensusArtifact
Validated consensus artifact.