Trait ic_interfaces::consensus_pool::ChangeSetOperation

pub trait ChangeSetOperation: Sized {
    fn or_else<F: FnOnce() -> Self>(self, f: F) -> Self;

    fn dedup_push(&mut self, action: ChangeAction) -> Result<(), ChangeAction>;
}

A trait with common methods for change sets.

Required methods

fn or_else<F: FnOnce() -> Self>(self, f: F) -> Self

Conditional composition when self is empty. Similar to Option::or_else.

fn dedup_push(&mut self, action: ChangeAction) -> Result<(), ChangeAction>

Append a change action only when it is not a duplicate of what already exists in the ChangeSet. Return the rejected action as error when it is considered as duplicate.

Loading content...