Trait ic_interfaces::consensus_pool::ConsensusPool

pub trait ConsensusPool {
    fn validated(&self) -> &dyn PoolSection<ValidatedConsensusArtifact>;

    fn unvalidated(&self) -> &dyn PoolSection<UnvalidatedConsensusArtifact>;

    fn as_cache(&self) -> &dyn ConsensusPoolCache;
}

The consensus pool contains all the artifacts received by P2P and produced by the local node.

It contains two sections:

The validated section contains artifacts that have been validated by consensus. To support resumability this section must be persistent.

The unvalidated section contains artifacts that have been received but haven't yet been validated. This section is in-memory only and thus volatile.

Required methods

fn validated(&self) -> &dyn PoolSection<ValidatedConsensusArtifact>

Return a reference to the validated PoolSection.

fn unvalidated(&self) -> &dyn PoolSection<UnvalidatedConsensusArtifact>

Return a reference to the unvalidated PoolSection.

fn as_cache(&self) -> &dyn ConsensusPoolCache

Return a reference to the consensus cache (ConsensusPoolCache).

Loading content...