Trait ic_interfaces::consensus_pool::ConsensusPoolCache

pub trait ConsensusPoolCache: Send + Sync {
    fn finalized_block(&self) -> Block;

    fn consensus_time(&self) -> Option<Time>;

    fn cup_with_protobuf(&self) -> CUPWithOriginalProtobuf;

    fn summary_block(&self) -> Block;

    fn catch_up_package(&self) -> CatchUpPackage { ... }
}

Reader of consensus related states.

Required methods

fn finalized_block(&self) -> Block

Return the latest/highest finalized block.

fn consensus_time(&self) -> Option<Time>

Return the time as recorded in the latest/highest finalized block. Return None if there has not been any finalized block since genesis.

fn cup_with_protobuf(&self) -> CUPWithOriginalProtobuf

Return the latest/highest CatchUpPackage together with its original protobuf.

fn summary_block(&self) -> Block

Return the latest/highest finalized block with DKG summary. In a situation where we have only finalized the catch-up block but not yet made a catch-up package, this will be different than the block in the latest catch-up package.

Loading content...
Provided methods

fn catch_up_package(&self) -> CatchUpPackage

Return the latest/highest CatchUpPackage.

Loading content...