Trait ic_interfaces::consensus_pool::HeightIndexedPool

pub trait HeightIndexedPool<T> {
    fn height_range(&self) -> Option<HeightRange>;

    fn max_height(&self) -> Option<Height>;

    fn get_all(&self) -> Box<dyn Iterator<Item = T>>;

    fn get_by_height(&self, h: Height) -> Box<dyn Iterator<Item = T>>;

    fn get_by_height_range(
        &self, 
        range: HeightRange
    ) -> Box<dyn Iterator<Item = T>>;

    fn get_only_by_height(&self, h: Height) -> Result<T, OnlyError>;

    fn get_highest(&self) -> Result<T, OnlyError>;

    fn get_highest_iter(&self) -> Box<dyn Iterator<Item = T>>;
}

HeightIndexedPool provides a set of interfaces for the Consensus component to query artifacts. The same interface is applicable to both validated and unvalidated partitions of consensus artifacts in the overall ArtifactPool.

Required methods

fn height_range(&self) -> Option<HeightRange>

Returns the height range of artifacts of type T currently in the pool.

fn max_height(&self) -> Option<Height>

Returns the max height across all artifacts of type T currently in the pool.

fn get_all(&self) -> Box<dyn Iterator<Item = T>>

Return an iterator over all of the artifacts of type T.

fn get_by_height(&self, h: Height) -> Box<dyn Iterator<Item = T>>

Return an iterator over the artifacts of type T at height 'h'.

fn get_by_height_range(&self, range: HeightRange) -> Box<dyn Iterator<Item = T>>

Return an iterator over the artifacts of type T in range range.min, range.max, inclusive. The items must be sorted by height in ascending order.

fn get_only_by_height(&self, h: Height) -> Result<T, OnlyError>

Return a single instance of artifact of type T, at height 'h', returning an error if there isn't one, or if there are more than one.

fn get_highest(&self) -> Result<T, OnlyError>

Return a single instance of artifact of type T at the highest height currently in the pool. Returns an error if there isn't one, or if there are more than one.

fn get_highest_iter(&self) -> Box<dyn Iterator<Item = T>>

Return an iterator over instances of artifact of type T at the highest height currently in the pool. Returns an error if there isn't one, or if there are more than one.

Loading content...