Trait ic_interfaces::consensus_pool::PoolSection

pub trait PoolSection<T> {
    fn contains(&self, msg_id: &ConsensusMessageId) -> bool;

    fn get(&self, msg_id: &ConsensusMessageId) -> Option<ConsensusMessage>;

    fn get_timestamp(&self, msg_id: &ConsensusMessageId) -> Option<Time>;

    fn random_beacon(&self) -> &dyn HeightIndexedPool<RandomBeacon>;

    fn block_proposal(&self) -> &dyn HeightIndexedPool<BlockProposal>;

    fn notarization(&self) -> &dyn HeightIndexedPool<Notarization>;

    fn finalization(&self) -> &dyn HeightIndexedPool<Finalization>;

    fn random_beacon_share(&self) -> &dyn HeightIndexedPool<RandomBeaconShare>;

    fn notarization_share(&self) -> &dyn HeightIndexedPool<NotarizationShare>;

    fn finalization_share(&self) -> &dyn HeightIndexedPool<FinalizationShare>;

    fn random_tape(&self) -> &dyn HeightIndexedPool<RandomTape>;

    fn random_tape_share(&self) -> &dyn HeightIndexedPool<RandomTapeShare>;

    fn catch_up_package(&self) -> &dyn HeightIndexedPool<CatchUpPackage>;

    fn catch_up_package_share(
        &self
    ) -> &dyn HeightIndexedPool<CatchUpPackageShare>;

    fn size(&self) -> u64;

    fn highest_catch_up_package_proto(&self) -> CatchUpPackage { ... }
}

A Pool section is a part of the consensus pool which contains artifacts.

Artifacts in the pool are accessible by their hash or by their type and height.

Required methods

fn contains(&self, msg_id: &ConsensusMessageId) -> bool

Checks if the artifact with the given Id is present in the pool

fn get(&self, msg_id: &ConsensusMessageId) -> Option<ConsensusMessage>

Lookup an artifact by ConsensusMessageId. Return the consensus message if it exists, or None otherwise.

fn get_timestamp(&self, msg_id: &ConsensusMessageId) -> Option<Time>

Lookup the timestamp of an artifact by its ConsensusMessageId.

fn random_beacon(&self) -> &dyn HeightIndexedPool<RandomBeacon>

Return the HeightIndexedPool for RandomBeacon.

fn block_proposal(&self) -> &dyn HeightIndexedPool<BlockProposal>

Return the HeightIndexedPool for BlockProposal.

fn notarization(&self) -> &dyn HeightIndexedPool<Notarization>

Return the HeightIndexedPool for Notarization.

fn finalization(&self) -> &dyn HeightIndexedPool<Finalization>

Return the HeightIndexedPool for Finalization.

fn random_beacon_share(&self) -> &dyn HeightIndexedPool<RandomBeaconShare>

Return the HeightIndexedPool for RandomBeaconShare.

fn notarization_share(&self) -> &dyn HeightIndexedPool<NotarizationShare>

Return the HeightIndexedPool for NotarizationShare.

fn finalization_share(&self) -> &dyn HeightIndexedPool<FinalizationShare>

Return the HeightIndexedPool for FinalizationShare.

fn random_tape(&self) -> &dyn HeightIndexedPool<RandomTape>

Return the HeightIndexedPool for RandomTape.

fn random_tape_share(&self) -> &dyn HeightIndexedPool<RandomTapeShare>

Return the HeightIndexedPool for RandomTapeShare.

fn catch_up_package(&self) -> &dyn HeightIndexedPool<CatchUpPackage>

Return the HeightIndexedPool for CatchUpPackage.

fn catch_up_package_share(&self) -> &dyn HeightIndexedPool<CatchUpPackageShare>

Return the HeightIndexedPool for CatchUpPackageShare.

fn size(&self) -> u64

Loading content...
Provided methods

fn highest_catch_up_package_proto(&self) -> CatchUpPackage

Return the HeightIndexedPool for CatchUpPackage in protobuf form.

Loading content...