Module ic_interfaces::crypto

The crypto public interface.

Structs

CryptoHashableTestDummy
A helper struct for testing that implements CryptoHashable.

SignableMock
A helper struct for testing that implements Signable.

Constants

DOMAIN_IC_REQUEST
The domain separator to be used when calculating the sender signature for a request to the Internet Computer according to the interface specification.

Traits

BasicSigVerifier
A Crypto Component interface to verify basic signatures.

BasicSigVerifierByPublicKey
A Crypto Component interface to verify basic signatures by public key.

BasicSigner
A Crypto Component interface to create basic signatures.

CanisterSigVerifier
A Crypto Component interface to verify (ICCSA) canister signatures.

Crypto
The functionality offered by the crypto component

CryptoHashDomain
A type that specifies a domain for a cryptographic hash.

CryptoHashable
A cryptographically hashable type.

DkgAlgorithm
The building blocks to perform distributed key generation (DKG).

ErrorReplication
A classifier for errors returned by the crypto component. Indicates whether a given error is permanent and guaranteed to occur in all replicas.

IngressSigVerifier
A Crypto Component interface to verify ingress messages.

KeyManager
Methods for checking and retrieving key material.

Keygen
Functions to generate key material.

MultiSigVerifier
A Crypto Component interface to verify and combine multi-signatures.

MultiSigner
A Crypto Component interface to create multi-signatures.

NiDkgAlgorithm
The building blocks to perform non-interactive distributed key generation (DKG).

Signable
Signable represents an object whose byte-vector representation can be signed using a digital signature scheme. It supports domain separation via SignatureDomain trait.

ThresholdSigVerifier
A Crypto Component interface to verify threshold signatures.

ThresholdSigVerifierByPublicKey
A Crypto Component interface to verify threshold signatures by a subnet's public key.

ThresholdSigner
A Crypto Component interface to create threshold signature shares.