Trait ic_interfaces::crypto::CryptoHashDomain

pub trait CryptoHashDomain: CryptoHashDomainSeal {
    fn domain(&self) -> String;
}

A type that specifies a domain for a cryptographic hash.

This trait is sealed and can only be implemented by types that are explicitly approved by the Github owners of this file (that is, the crypto team) via an implementation of the CryptoHashDomainSeal. Explicit approval is required for security reasons to ensure proper domain separation.

Required methods

fn domain(&self) -> String

Returns the domain separator used in cryptographic hashes.

Loading content...