Trait ic_interfaces::crypto::DkgAlgorithm

pub trait DkgAlgorithm {
    fn generate_encryption_keys(
        &self, 
        config: &Config, 
        node_id: NodeId
    ) -> CryptoResult<EncryptionPublicKeyWithPop>;

    fn verify_encryption_public_key(
        &self, 
        config: &Config, 
        sender: NodeId, 
        key: &EncryptionPublicKeyWithPop
    ) -> CryptoResult<()>;

    fn create_dealing(
        &self, 
        config: &Config, 
        verified_keys: &BTreeMap<NodeId, EncryptionPublicKeyWithPop>, 
        node_id: NodeId
    ) -> CryptoResult<Dealing>;

    fn verify_dealing(
        &self, 
        config: &Config, 
        verified_keys: &BTreeMap<NodeId, EncryptionPublicKeyWithPop>, 
        dealer: NodeId, 
        dealing: &Dealing
    ) -> CryptoResult<()>;

    fn create_response(
        &self, 
        config: &Config, 
        verified_keys: &BTreeMap<NodeId, EncryptionPublicKeyWithPop>, 
        verified_dealings: &BTreeMap<NodeId, Dealing>, 
        node_id: NodeId
    ) -> CryptoResult<Response>;

    fn verify_response(
        &self, 
        config: &Config, 
        verified_keys: &BTreeMap<NodeId, EncryptionPublicKeyWithPop>, 
        verified_dealings: &BTreeMap<NodeId, Dealing>, 
        receiver: NodeId, 
        response: &Response
    ) -> CryptoResult<()>;

    fn create_transcript(
        &self, 
        config: &Config, 
        verified_keys: &BTreeMap<NodeId, EncryptionPublicKeyWithPop>, 
        verified_dealings: &BTreeMap<NodeId, Dealing>, 
        verified_responses: &BTreeMap<NodeId, Response>
    ) -> CryptoResult<Transcript>;

    fn load_transcript(
        &self, 
        transcript: &Transcript, 
        node_id: NodeId
    ) -> CryptoResult<()>;
}

The building blocks to perform distributed key generation (DKG).

There are two variants of DKG that are supported:

(Non-Resharing) DKG

In the standard DKG, the group public key changes in every DKG phase.

To run this variant of DKG, always use a dkg::Config that has resharing_transcript = None.

Resharing DKG

In resharing DKG, the group public key remains the same, as opposed to changing in every DKG phase.

To run non-resharing DKG, proceed as follows:

In the first phase, run (non-resharing) DKG using a dkg::Config that has resharing_transcript = None.

With the resulting Transcript, re-sharing DKG can be run repeatedly by using a dkg::Config that has resharing_transcript = Some(transcript), where transcript is the Transcript from the previous DKG phase.

Required methods

fn generate_encryption_keys(
&self,
config: &Config,
node_id: NodeId
) -> CryptoResult<EncryptionPublicKeyWithPop>

Generate an (ephemeral) encryption key pair, store the secret key in the local store, and return the public key.

fn verify_encryption_public_key(
&self,
config: &Config,
sender: NodeId,
key: &EncryptionPublicKeyWithPop
) -> CryptoResult<()>

Verify an (ephemeral) public encryption key and its proof of possession.

fn create_dealing(
&self,
config: &Config,
verified_keys: &BTreeMap<NodeId, EncryptionPublicKeyWithPop>,
node_id: NodeId
) -> CryptoResult<Dealing>

Create a dealing.

Preconditions

for each (node_id, pk) in verified_keys, it holds that verify_encryption_public_key(config, node_id, pk).is_ok()

fn verify_dealing(
&self,
config: &Config,
verified_keys: &BTreeMap<NodeId, EncryptionPublicKeyWithPop>,
dealer: NodeId,
dealing: &Dealing
) -> CryptoResult<()>

Verify a dealing.

Preconditions

for each (node_id, pk) in verified_keys, it holds that verify_encryption_public_key(config, node_id, pk).is_ok()

fn create_response(
&self,
config: &Config,
verified_keys: &BTreeMap<NodeId, EncryptionPublicKeyWithPop>,
verified_dealings: &BTreeMap<NodeId, Dealing>,
node_id: NodeId
) -> CryptoResult<Response>

Create a response.

Preconditions

for each (node_id, pk) in verified_keys, it holds that verify_encryption_public_key(config, node_id, pk).is_ok()

for each (dealer, dealing) in verified_dealings, it holds that verify_dealing(config, verified_keys, dealer, dealing).is_ok()

fn verify_response(
&self,
config: &Config,
verified_keys: &BTreeMap<NodeId, EncryptionPublicKeyWithPop>,
verified_dealings: &BTreeMap<NodeId, Dealing>,
receiver: NodeId,
response: &Response
) -> CryptoResult<()>

Verify a response.

Preconditions

for each (node_id, pk) in verified_keys, it holds that verify_encryption_public_key(config, node_id, pk).is_ok()

for each (dealer, dealing) in verified_dealings, it holds that verify_dealing(config, verified_keys, dealer, dealing).is_ok()

fn create_transcript(
&self,
config: &Config,
verified_keys: &BTreeMap<NodeId, EncryptionPublicKeyWithPop>,
verified_dealings: &BTreeMap<NodeId, Dealing>,
verified_responses: &BTreeMap<NodeId, Response>
) -> CryptoResult<Transcript>

Combine the given dealings into a transcript.

Preconditions

for each (node_id, pk) in verified_keys, it holds that verify_encryption_public_key(config, node_id, pk).is_ok()

for each (dealer, dealing) in verified_dealings, it holds that verify_dealing(config, verified_keys, dealer, dealing).is_ok()

for each (receiver, response) in verified_responses, it holds that verify_response(config, verified_keys, verified_dealings, receiver, response).is_ok()

fn load_transcript(
&self,
transcript: &Transcript,
node_id: NodeId
) -> CryptoResult<()>

Load the given transcript. This ensures that

the (group's) public threshold signing key is available for the methods of the ThresholdSigVerifier trait, and

the individual nodes' public threshold signing keys are available for the methods of the ThresholdSigVerifier trait, and

if the node is a receiver, the node's secret threshold signing key is available for the methods of the ThresholdSigner trait for the DKG instance identified by the DKG ID contained in the transcript.

Loading content...