Trait ic_interfaces::crypto::ErrorReplication

pub trait ErrorReplication {
    fn is_replicated(&self) -> bool;
}

A classifier for errors returned by the crypto component. Indicates whether a given error is permanent and guaranteed to occur in all replicas.

Required methods

fn is_replicated(&self) -> bool

Loading content...