Trait ic_interfaces::crypto::KeyManager

pub trait KeyManager {
    fn check_keys_with_registry(
        &self, 
        registry_version: RegistryVersion
    ) -> CryptoResult<()>;

    fn node_public_keys(&self) -> NodePublicKeys;
}

Methods for checking and retrieving key material.

Required methods

fn check_keys_with_registry(
&self,
registry_version: RegistryVersion
) -> CryptoResult<()>

Checks whether this crypto component is properly set up, i.e. whether the registry contains the required public keys, and whether the crypto component's secret key store contains the corresponding secret keys.

fn node_public_keys(&self) -> NodePublicKeys

Returns node public keys that were read when this crypto component was created. Node public keys stay the same throughout the lifetime of the component.

Loading content...