Trait ic_interfaces::crypto::ThresholdSigVerifier

pub trait ThresholdSigVerifier<T: Signable> {
    fn verify_threshold_sig_share(
        &self, 
        signature: &ThresholdSigShareOf<T>, 
        message: &T, 
        dkg_id: DkgId, 
        signer: NodeId
    ) -> CryptoResult<()>;

    fn combine_threshold_sig_shares(
        &self, 
        shares: BTreeMap<NodeId, ThresholdSigShareOf<T>>, 
        dkg_id: DkgId
    ) -> CryptoResult<CombinedThresholdSigOf<T>>;

    fn verify_threshold_sig_combined(
        &self, 
        signature: &CombinedThresholdSigOf<T>, 
        message: &T, 
        dkg_id: DkgId
    ) -> CryptoResult<()>;
}

A Crypto Component interface to verify threshold signatures.

The methods in this trait depend on key material for the respective DKG ID to be present in the cache of the crypto component. To initialize this key material, the DKG transcript for the respective DKG ID must be loaded by using the load_transcript method of the DkgAlgorithm trait (which is implemented by the crypto component).

Required methods

fn verify_threshold_sig_share(
&self,
signature: &ThresholdSigShareOf<T>,
message: &T,
dkg_id: DkgId,
signer: NodeId
) -> CryptoResult<()>

Verifies a threshold signature share.

See the trait's doc comment for applicable preconditions.

Errors

CryptoError::SignatureVerification if the signature is not valid

CryptoError::ThresholdSigDataNotFound if the threshold signature data store does not contain the necessary information for the dkg_id to calculate the individual public key required for verifying the signature. This error indicates that DkgAlgorithm::load_transcript must be called prior to calling this method.

CryptoError::MalformedSignature if the given signature is not a valid threshold signature (e.g., because of invalid length).

CryptoError::InvalidArgument if the transcript data in the threshold signature data store does not contain a node index for the given signer. This indicates that the signer is not a valid receiver for the DKG round with ID dkg_id.

Panics

This method panics if calculating the individual public key fails, which would happen if the implementations of DKG and threshold signatures are not aligned.

This method panics if verifying the signature fails because the implementations of DKG and threshold signatures are not aligned.

fn combine_threshold_sig_shares(
&self,
shares: BTreeMap<NodeId, ThresholdSigShareOf<T>>,
dkg_id: DkgId
) -> CryptoResult<CombinedThresholdSigOf<T>>

Combines the given threshold signature shares.

See the trait's doc comment for applicable preconditions.

Errors

CryptoError::InvalidArgument if the shares are empty, the threshold is too high, no index is present in the store for a node id in shares, the algorithm is unknown, or the public coefficients cannot be parsed.

CryptoError::ThresholdSigDataNotFound if the threshold signature data store does not contain the public coefficients and node indices for the dkg_id. This error indicates that DkgAlgorithm::load_transcript must be called prior to calling this method.

CryptoError::MalformedSignature if a sig share is malformed.

fn verify_threshold_sig_combined(
&self,
signature: &CombinedThresholdSigOf<T>,
message: &T,
dkg_id: DkgId
) -> CryptoResult<()>

Verifies a combined threshold signature.

See the trait's doc comment for applicable preconditions.

Errors

CryptoError::SignatureVerification if the signature is not valid

CryptoError::ThresholdSigDataNotFound if the threshold signature data store does not contain the necessary information for the dkg_id. This error indicates that DkgAlgorithm::load_transcript must be called prior to calling this method.

CryptoError::MalformedSignature if the given signature is not a valid combined threshold signature (e.g., because of invalid length)

Loading content...