Trait ic_interfaces::crypto::ThresholdSigner

pub trait ThresholdSigner<T: Signable> {
    fn sign_threshold(
        &self, 
        message: &T, 
        dkg_id: DkgId
    ) -> CryptoResult<ThresholdSigShareOf<T>>;
}

A Crypto Component interface to create threshold signature shares.

Required methods

fn sign_threshold(
&self,
message: &T,
dkg_id: DkgId
) -> CryptoResult<ThresholdSigShareOf<T>>

Creates a threshold signature share.

This method depends on key material for the respective DKG ID to be present in the secret key store (SKS) of the crypto component. To initialize this key material, the DKG transcript for the respective DKG ID must be loaded by using the load_transcript method of the DkgAlgorithm trait, which is implemented by the crypto component.

Errors

CryptoError::ThresholdSigDataNotFound if the threshold signature data store does not contain public coefficients for the dkg_id. This error indicates that NiDkgAlgorithm::load_transcript (or DkgAlgorithm::load_transcript, respectively) must be called prior to calling this method.

CryptoError::SecretKeyNotFound if the secret key is not present in the secret key store.

Loading content...