Trait ic_interfaces::dkg::DkgPool

pub trait DkgPool: Send + Sync {
    fn get_validated(&self) -> Box<dyn Iterator<Item = &Message>>;

    fn get_validated_older_than(
        &self, 
        age_threshold: Duration
    ) -> Box<dyn Iterator<Item = &Message>>;

    fn get_unvalidated(&self) -> Box<dyn Iterator<Item = &Message>>;

    fn get_current_start_height(&self) -> Height;

    fn validated_contains(&self, msg: &Message) -> bool;
}

The DkgPool is used to store messages that are exchanged between nodes in the process of executing dkg.

Required methods

fn get_validated(&self) -> Box<dyn Iterator<Item = &Message>>

fn get_validated_older_than(
&self,
age_threshold: Duration
) -> Box<dyn Iterator<Item = &Message>>

Returns the validated entries older than the age threshold

fn get_unvalidated(&self) -> Box<dyn Iterator<Item = &Message>>

fn get_current_start_height(&self) -> Height

The start height of the currently computed DKG interval; the invariant we want to maintain for all messages in validated and unvalidated sections is that they correspond to a DKG Id with the start height equal to current_start_height.

fn validated_contains(&self, msg: &Message) -> bool

Checks if the message is present in the validated section.

Loading content...