Trait ic_interfaces::execution_environment::ExecutionEnvironment

pub trait ExecutionEnvironment: Sync + Send {
    type State;
    type CanisterState;
    fn execute_subnet_message(
        &self, 
        msg: CanisterInputMessage, 
        state: Self::State, 
        instructions_limit: NumInstructions, 
        rng: &mut (dyn RngCore + 'static), 
        provisional_whitelist: &ProvisionalWhitelist, 
        subnet_available_memory: SubnetAvailableMemory
    ) -> Self::State;

    fn execute_canister_message(
        &self, 
        canister_state: Self::CanisterState, 
        instructions_limit: NumInstructions, 
        msg: CanisterInputMessage, 
        time: Time, 
        routing_table: Arc<RoutingTable>, 
        subnet_records: Arc<BTreeMap<SubnetId, SubnetType>>, 
        subnet_available_memory: SubnetAvailableMemory
    ) -> ExecResult<ExecuteMessageResult<Self::CanisterState>>;

    fn should_accept_ingress_message(
        &self, 
        state: Arc<Self::State>, 
        provisional_whitelist: &ProvisionalWhitelist, 
        ingress: &SignedIngressContent
    ) -> Result<(), MessageAcceptanceError>;

    fn execute_canister_heartbeat(
        &self, 
        canister_state: Self::CanisterState, 
        instructions_limit: NumInstructions, 
        routing_table: Arc<RoutingTable>, 
        subnet_records: Arc<BTreeMap<SubnetId, SubnetType>>, 
        time: Time, 
        subnet_available_memory: SubnetAvailableMemory
    ) -> ExecResult<(Self::CanisterState, NumInstructions, Result<NumBytes, CanisterHeartbeatError>)>;

    fn subnet_available_memory(&self, state: &Self::State) -> NumBytes;
}

ExecutionEnvironment is the component responsible for executing messages on the IC.

Associated Types

type State

Type modelling the replicated state.

Should typically be ic_replicated_state::ReplicatedState.

type CanisterState

Type modelling the canister state.

Should typically be ic_replicated_state::CanisterState.

Loading content...
Required methods

fn execute_subnet_message(
&self,
msg: CanisterInputMessage,
state: Self::State,
instructions_limit: NumInstructions,
rng: &mut (dyn RngCore + 'static),
provisional_whitelist: &ProvisionalWhitelist,
subnet_available_memory: SubnetAvailableMemory
) -> Self::State

Executes a message sent to a subnet.

fn execute_canister_message(
&self,
canister_state: Self::CanisterState,
instructions_limit: NumInstructions,
msg: CanisterInputMessage,
time: Time,
routing_table: Arc<RoutingTable>,
subnet_records: Arc<BTreeMap<SubnetId, SubnetType>>,
subnet_available_memory: SubnetAvailableMemory
) -> ExecResult<ExecuteMessageResult<Self::CanisterState>>

Executes a message sent to a canister.

fn should_accept_ingress_message(
&self,
state: Arc<Self::State>,
provisional_whitelist: &ProvisionalWhitelist,
ingress: &SignedIngressContent
) -> Result<(), MessageAcceptanceError>

Asks the canister if it is willing to accept the provided ingress message.

fn execute_canister_heartbeat(
&self,
canister_state: Self::CanisterState,
instructions_limit: NumInstructions,
routing_table: Arc<RoutingTable>,
subnet_records: Arc<BTreeMap<SubnetId, SubnetType>>,
time: Time,
subnet_available_memory: SubnetAvailableMemory
) -> ExecResult<(Self::CanisterState, NumInstructions, Result<NumBytes, CanisterHeartbeatError>)>

Executes a heartbeat of a given canister.

fn subnet_available_memory(&self, state: &Self::State) -> NumBytes

Look up the current amount of memory available on the subnet. EXC-185 will make this method obsolete.

Loading content...