Trait ic_interfaces::execution_environment::IngressHistoryReader

pub trait IngressHistoryReader: Send + Sync {
    fn get_latest_status(&self) -> Box<dyn Fn(&MessageId) -> IngressStatus>;

    fn get_status_at_height(
        &self, 
        height: Height
    ) -> Result<Box<dyn Fn(&MessageId) -> IngressStatus>, IngressHistoryError>;
}

Interface for reading the history of ingress messages.

Required methods

fn get_latest_status(&self) -> Box<dyn Fn(&MessageId) -> IngressStatus>

Returns a function that can be used to query the status for a given message_id using the latest execution state.

fn get_status_at_height(
&self,
height: Height
) -> Result<Box<dyn Fn(&MessageId) -> IngressStatus>, IngressHistoryError>

Return a function that can be used to query the status for a given message_id using the state at given height.

Return an error if the the state is not available.

Loading content...