Trait ic_interfaces::execution_environment::SystemApi

pub trait SystemApi {
    fn set_execution_error(&mut self, error: HypervisorError);

    fn get_execution_error(&self) -> Option<&HypervisorError>;

    fn get_available_num_instructions(&self) -> NumInstructions;

    fn get_stable_memory_delta_pages(&self) -> usize;

    fn set_available_num_instructions(
        &mut self, 
        num_instructions: NumInstructions
    );

    fn ic0_msg_caller_copy(
        &self, 
        dst: u32, 
        offset: u32, 
        size: u32, 
        heap: &mut [u8]
    ) -> HypervisorResult<()>;

    fn ic0_msg_caller_size(&self) -> HypervisorResult<u32>;

    fn ic0_msg_arg_data_size(&self) -> HypervisorResult<u32>;

    fn ic0_msg_arg_data_copy(
        &self, 
        dst: u32, 
        offset: u32, 
        size: u32, 
        heap: &mut [u8]
    ) -> HypervisorResult<()>;

    fn ic0_msg_method_name_size(&self) -> HypervisorResult<u32>;

    fn ic0_msg_method_name_copy(
        &self, 
        dst: u32, 
        offset: u32, 
        size: u32, 
        heap: &mut [u8]
    ) -> HypervisorResult<()>;

    fn ic0_accept_message(&mut self) -> HypervisorResult<()>;

    fn ic0_msg_reply_data_append(
        &mut self, 
        src: u32, 
        size: u32, 
        heap: &[u8]
    ) -> HypervisorResult<()>;

    fn ic0_msg_reply(&mut self) -> HypervisorResult<()>;

    fn ic0_msg_reject_code(&self) -> HypervisorResult<i32>;

    fn ic0_msg_reject(
        &mut self, 
        src: u32, 
        size: u32, 
        heap: &[u8]
    ) -> HypervisorResult<()>;

    fn ic0_msg_reject_msg_size(&self) -> HypervisorResult<u32>;

    fn ic0_msg_reject_msg_copy(
        &self, 
        dst: u32, 
        offset: u32, 
        size: u32, 
        heap: &mut [u8]
    ) -> HypervisorResult<()>;

    fn ic0_canister_self_size(&self) -> HypervisorResult<usize>;

    fn ic0_canister_self_copy(
        &mut self, 
        dst: u32, 
        offset: u32, 
        size: u32, 
        heap: &mut [u8]
    ) -> HypervisorResult<()>;

    fn ic0_controller_size(&self) -> HypervisorResult<usize>;

    fn ic0_controller_copy(
        &mut self, 
        dst: u32, 
        offset: u32, 
        size: u32, 
        heap: &mut [u8]
    ) -> HypervisorResult<()>;

    fn ic0_debug_print(&self, src: u32, size: u32, heap: &[u8]);

    fn ic0_exec(&mut self, bytes: Vec<u8>, payload: Vec<u8>) -> HypervisorError;

    fn ic0_trap(&self, src: u32, size: u32, heap: &[u8]) -> HypervisorError;

    fn ic0_call_simple(
        &mut self, 
        callee_src: u32, 
        callee_size: u32, 
        method_name_src: u32, 
        method_name_len: u32, 
        reply_fun: u32, 
        reply_env: u32, 
        reject_fun: u32, 
        reject_env: u32, 
        data_src: u32, 
        data_len: u32, 
        heap: &[u8]
    ) -> HypervisorResult<i32>;

    fn ic0_call_new(
        &mut self, 
        callee_src: u32, 
        callee_size: u32, 
        name_src: u32, 
        name_len: u32, 
        reply_fun: u32, 
        reply_env: u32, 
        reject_fun: u32, 
        reject_env: u32, 
        heap: &[u8]
    ) -> HypervisorResult<()>;

    fn ic0_call_data_append(
        &mut self, 
        src: u32, 
        size: u32, 
        heap: &[u8]
    ) -> HypervisorResult<()>;

    fn ic0_call_on_cleanup(
        &mut self, 
        fun: u32, 
        env: u32
    ) -> HypervisorResult<()>;

    fn ic0_call_cycles_add(&mut self, amount: u64) -> HypervisorResult<()>;

    fn ic0_call_perform(&mut self) -> HypervisorResult<i32>;

    fn ic0_stable_size(&self) -> HypervisorResult<u32>;

    fn ic0_stable_grow(
        &mut self, 
        additional_pages: u32
    ) -> HypervisorResult<i32>;

    fn ic0_stable_read(
        &self, 
        dst: u32, 
        offset: u32, 
        size: u32, 
        heap: &mut [u8]
    ) -> HypervisorResult<()>;

    fn ic0_stable_write(
        &mut self, 
        offset: u32, 
        src: u32, 
        size: u32, 
        heap: &[u8]
    ) -> HypervisorResult<()>;

    fn ic0_time(&self) -> HypervisorResult<Time>;

    fn out_of_instructions(&self) -> HypervisorResult<NumInstructions>;

    fn update_available_memory(
        &mut self, 
        native_memory_grow_res: i32, 
        additional_pages: u32
    ) -> HypervisorResult<i32>;

    fn ic0_canister_cycle_balance(&self) -> HypervisorResult<u64>;

    fn ic0_msg_cycles_available(&self) -> HypervisorResult<u64>;

    fn ic0_msg_cycles_refunded(&self) -> HypervisorResult<u64>;

    fn ic0_msg_cycles_accept(
        &mut self, 
        max_amount: u64
    ) -> HypervisorResult<u64>;

    fn ic0_certified_data_set(
        &mut self, 
        src: u32, 
        size: u32, 
        heap: &[u8]
    ) -> HypervisorResult<()>;

    fn ic0_data_certificate_present(&self) -> HypervisorResult<i32>;

    fn ic0_data_certificate_size(&self) -> HypervisorResult<i32>;

    fn ic0_data_certificate_copy(
        &self, 
        dst: u32, 
        offset: u32, 
        size: u32, 
        heap: &mut [u8]
    ) -> HypervisorResult<()>;

    fn ic0_canister_status(&self) -> HypervisorResult<u32>;

    fn ic0_mint_cycles(&mut self, amount: u64) -> HypervisorResult<u64>;
}

A trait for providing all necessary imports to a Wasm module.

Required methods

fn set_execution_error(&mut self, error: HypervisorError)

Stores the execution error, so that the user can evaluate it later.

fn get_execution_error(&self) -> Option<&HypervisorError>

Returns the reference to the execution error.

fn get_available_num_instructions(&self) -> NumInstructions

Returns the amount of available instructions.

fn get_stable_memory_delta_pages(&self) -> usize

Returns the stable memory delta that the canister produced

fn set_available_num_instructions(&mut self, num_instructions: NumInstructions)

Sets the amount of available instructions.

fn ic0_msg_caller_copy(
&self,
dst: u32,
offset: u32,
size: u32,
heap: &mut [u8]
) -> HypervisorResult<()>

Copies size bytes starting from offset inside the opaque caller blob and copies them to heap[dst..dst+size]. The caller is the canister id in case of requests or the user id in case of an ingress message.

fn ic0_msg_caller_size(&self) -> HypervisorResult<u32>

Returns the size of the opaque caller blob.

fn ic0_msg_arg_data_size(&self) -> HypervisorResult<u32>

Returns the size of msg.payload.

fn ic0_msg_arg_data_copy(
&self,
dst: u32,
offset: u32,
size: u32,
heap: &mut [u8]
) -> HypervisorResult<()>

Copies length bytes from msg.payload[offset..offset+size] to memory[dst..dst+size].

fn ic0_msg_method_name_size(&self) -> HypervisorResult<u32>

Used to look up the size of the method_name that the message wants to call. Can only be called in the context of inspecting messages.

fn ic0_msg_method_name_copy(
&self,
dst: u32,
offset: u32,
size: u32,
heap: &mut [u8]
) -> HypervisorResult<()>

Used to copy the method_name that the message wants to call to heap. Can only be called in the context of inspecting messages.

fn ic0_accept_message(&mut self) -> HypervisorResult<()>

fn ic0_msg_reply_data_append(
&mut self,
src: u32,
size: u32,
heap: &[u8]
) -> HypervisorResult<()>

Copies the data referred to by src/size out of the canister and appends it to the (initially empty) data reply.

fn ic0_msg_reply(&mut self) -> HypervisorResult<()>

Replies to the sender with the data assembled using msg_reply_data_append.

fn ic0_msg_reject_code(&self) -> HypervisorResult<i32>

Returns the reject code, if the current function is invoked as a reject callback.

It returns the special “no error” code 0 if the callback is not invoked as a reject callback

fn ic0_msg_reject(
&mut self,
src: u32,
size: u32,
heap: &[u8]
) -> HypervisorResult<()>

Replies to sender with an error message

fn ic0_msg_reject_msg_size(&self) -> HypervisorResult<u32>

Returns the length of the reject message in bytes.

Panics

This traps if not invoked from a reject callback.

fn ic0_msg_reject_msg_copy(
&self,
dst: u32,
offset: u32,
size: u32,
heap: &mut [u8]
) -> HypervisorResult<()>

Copies length bytes from self.reject_msg[offset..offset+size] to memory[dst..dst+size]

Panics

This traps if offset+size is greater than the size of the reject message, or if dst+size exceeds the size of the Wasm memory, or if not called from inside a reject callback.

fn ic0_canister_self_size(&self) -> HypervisorResult<usize>

Returns the size of the blob corresponding to the id of the canister.

fn ic0_canister_self_copy(
&mut self,
dst: u32,
offset: u32,
size: u32,
heap: &mut [u8]
) -> HypervisorResult<()>

Copies size bytes starting from offset in the id blob of the canister to heap[dst..dst+size].

fn ic0_controller_size(&self) -> HypervisorResult<usize>

Returns the size of the blob corresponding to the id of the controller.

fn ic0_controller_copy(
&mut self,
dst: u32,
offset: u32,
size: u32,
heap: &mut [u8]
) -> HypervisorResult<()>

Copies size bytes starting from offset in the id blob of the controller to heap[dst..dst+size].

fn ic0_debug_print(&self, src: u32, size: u32, heap: &[u8])

Outputs the specified bytes on the heap as a string on STDOUT.

fn ic0_exec(&mut self, bytes: Vec<u8>, payload: Vec<u8>) -> HypervisorError

Just like exec in C replaces the current process with a new process, this system call replaces the current canister with a new canister.

fn ic0_trap(&self, src: u32, size: u32, heap: &[u8]) -> HypervisorError

Traps, with a possibly helpful message

fn ic0_call_simple(
&mut self,
callee_src: u32,
callee_size: u32,
method_name_src: u32,
method_name_len: u32,
reply_fun: u32,
reply_env: u32,
reject_fun: u32,
reject_env: u32,
data_src: u32,
data_len: u32,
heap: &[u8]
) -> HypervisorResult<i32>

Creates a pending inter-canister message that will be scheduled if the current message execution completes successfully.

fn ic0_call_new(
&mut self,
callee_src: u32,
callee_size: u32,
name_src: u32,
name_len: u32,
reply_fun: u32,
reply_env: u32,
reject_fun: u32,
reject_env: u32,
heap: &[u8]
) -> HypervisorResult<()>

Begins assembling a call to the canister specified by callee_src/callee_size at method name_src/name_size. Two mandatory callbacks are recorded which will be invoked on success and error respectively.

Subsequent calls to other call_* apis set further attributes of this call until the call is concluded (with ic0.call_perform) or discarded (by returning without calling ic0.call_performor by starting a new call withic0.call_new`).

fn ic0_call_data_append(
&mut self,
src: u32,
size: u32,
heap: &[u8]
) -> HypervisorResult<()>

Appends the specified bytes to the argument of the call. Initially, the argument is empty. This can be called multiple times between ic0.call_new and ic0.call_perform.

fn ic0_call_on_cleanup(&mut self, fun: u32, env: u32) -> HypervisorResult<()>

Specifies the closure to be called if the reply/reject closures trap. Can be called at most once between ic0.call_new and ic0.call_perform.

See https://sdk.dfinity.org/docs/interface-spec/index.html#system-api-call

fn ic0_call_cycles_add(&mut self, amount: u64) -> HypervisorResult<()>

Adds cycles to a call by moving them from the canister's balance onto the call under construction. The cycles are deducted immediately from the canister's balance and moved back if the call cannot be performed (e.g. if ic0.call_perform signals an error or if the canister invokes ic0.call_new or returns without invoking ic0.call_perform).

This traps if trying to transfer more cycles than are in the current balance of the canister.

fn ic0_call_perform(&mut self) -> HypervisorResult<i32>

This call concludes assembling the call. It queues the call message to the given destination, but does not actually act on it until the current WebAssembly function returns without trapping.

If the system returns 0, then the system was able to enqueue the call, if a non-zero value is returned then the call could not be enqueued.

After ic0.call_perform and before the next ic0.call_new, all other ic0.call_* calls trap.

fn ic0_stable_size(&self) -> HypervisorResult<u32>

Returns the current size of the stable memory in WebAssembly pages.

fn ic0_stable_grow(&mut self, additional_pages: u32) -> HypervisorResult<i32>

Tries to grow the stable memory by additional_pages many pages containing zeros. If successful, returns the previous size of the memory (in pages). Otherwise, returns -1

fn ic0_stable_read(
&self,
dst: u32,
offset: u32,
size: u32,
heap: &mut [u8]
) -> HypervisorResult<()>

Copies the data referred to by offset/size out of the stable memory and replaces the corresponding bytes starting at dst in the canister memory.

This system call traps if dst+size exceeds the size of the WebAssembly memory or offset+size exceeds the size of the stable memory.

fn ic0_stable_write(
&mut self,
offset: u32,
src: u32,
size: u32,
heap: &[u8]
) -> HypervisorResult<()>

Copies the data referred to by src/size out of the canister and replaces the corresponding segment starting at offset in the stable memory.

This system call traps if src+size exceeds the size of the WebAssembly memory or offset+size exceeds the size of the stable memory.

fn ic0_time(&self) -> HypervisorResult<Time>

fn out_of_instructions(&self) -> HypervisorResult<NumInstructions>

This system call is not part of the public spec and used by the hypervisor, when execution runs out of instructions. Higher levels can decide how to proceed, by either providing more instructions or aborting the execution (typically with an out-of-instructions error).

fn update_available_memory(
&mut self,
native_memory_grow_res: i32,
additional_pages: u32
) -> HypervisorResult<i32>

This system call is not part of the public spec. It's called after a native memory.grow has been called to check whether there's enough available memory left.

fn ic0_canister_cycle_balance(&self) -> HypervisorResult<u64>

Returns the current balance in cycles.

fn ic0_msg_cycles_available(&self) -> HypervisorResult<u64>

Cycles sent in the current call and still available.

fn ic0_msg_cycles_refunded(&self) -> HypervisorResult<u64>

Cycles that came back with the response, as a refund.

fn ic0_msg_cycles_accept(&mut self, max_amount: u64) -> HypervisorResult<u64>

This moves cycles from the call to the canister balance. It can be called multiple times, each time adding more cycles to the balance.

It moves no more cycles than max_amount.

It moves no more cycles than available according to ic0.msg_cycles_available, and

The canister balance afterwards does not exceed maximum amount of cycles it can hold (public spec refers to this constant as MAX_CANISTER_BALANCE) minus any possible outstanding balances. However, canisters on system subnets have no balance limit.

EXE-117: the last point is not properly handled yet. In particular, a refund can come back to the canister after this call finishes which causes the canister's balance to overflow.

fn ic0_certified_data_set(
&mut self,
src: u32,
size: u32,
heap: &[u8]
) -> HypervisorResult<()>

Sets the certified data for the canister. See: https://sdk.dfinity.org/docs/interface-spec/index.html#system-api-certified-data

fn ic0_data_certificate_present(&self) -> HypervisorResult<i32>

If run in non-replicated execution (i.e. query), returns 1 if the data certificate is present, 0 otherwise. If run in replicated execution (i.e. an update call or a certified query), returns 0.

fn ic0_data_certificate_size(&self) -> HypervisorResult<i32>

Returns the size of the data certificate if it is present (i.e. data_certificate_present returns 1). Traps if data_certificate_present returns 0.

fn ic0_data_certificate_copy(
&self,
dst: u32,
offset: u32,
size: u32,
heap: &mut [u8]
) -> HypervisorResult<()>

Copies the data certificate into the heap if it is present (i.e. data_certificate_present returns 1). Traps if data_certificate_present returns 0.

fn ic0_canister_status(&self) -> HypervisorResult<u32>

Returns the current status of the canister. 1 indicates running, 2 indicates stopping, and 3 indicates stopped.

fn ic0_mint_cycles(&mut self, amount: u64) -> HypervisorResult<u64>

Mints the amount cycles Adds cycles to the canister's balance.

Adds no more cycles than amount.

The canister balance afterwards does not exceed maximum amount of cycles it can hold. However, canisters on system subnets have no balance limit.

Returns the amount of cycles added to the canister's balance.

Loading content...