Trait ic_interfaces::gossip_pool::GossipPool

pub trait GossipPool<T, S> {
    type MessageId;
    type Filter;
    fn contains(&self, id: &Self::MessageId) -> bool;

    fn get_validated_by_identifier(&self, id: &Self::MessageId) -> Option<T>;

    fn get_all_validated_by_filter(
        &self, 
        filter: Self::Filter
    ) -> Box<dyn Iterator<Item = T>>;

    fn check_quota(
        &self, 
        _msg: &T, 
        _peer_id: &NodeId
    ) -> Result<(), ArtifactPoolError> { ... }
}

GossipPool trait is the generic interface used by ArtifactManager to interact with the Pools internally and allow GossipProtocol to serve the gossip functionality. Every pool needs to implement this trait.

Associated Types

type MessageId

type Filter

Loading content...
Required methods

fn contains(&self, id: &Self::MessageId) -> bool

Check if an artifact exists by its Id.

fn get_validated_by_identifier(&self, id: &Self::MessageId) -> Option<T>

Get a validated artifact by its identifier

#Returns:

'Some`: Artifact from the validated pool.

None: Artifact does not exist in the validated pool.

fn get_all_validated_by_filter(
&self,
filter: Self::Filter
) -> Box<dyn Iterator<Item = T>>

Get all validated artifacts by the filter See interfaces/src/artifact_manager.rs for more details

#Returns: A iterator over all the validated artifacts.

Loading content...
Provided methods

fn check_quota(
&self,
_msg: &T,
_peer_id: &NodeId
) -> Result<(), ArtifactPoolError>

Check if an unvalidated artifact can be inserted into the pool #Returns:

Ok: The artifact can be inserted into the unvalidated pool.

Error: If there is insufficient quota available for the peer, Return InsufficientQuotaError.

Default implementation is just to return Ok.

Loading content...