Trait ic_interfaces::ingress_manager::IngressSelector

pub trait IngressSelector: Send + Sync {
    fn get_ingress_payload(
        &self, 
        ingress_pool: &dyn IngressPoolSelect, 
        past_ingress: &dyn IngressSetQuery, 
        context: &ValidationContext
    ) -> IngressPayload;

    fn validate_ingress_payload(
        &self, 
        payload: &IngressPayload, 
        past_ingress: &dyn IngressSetQuery, 
        context: &ValidationContext
    ) -> ValidationResult<IngressPayloadValidationError>;

    fn request_purge_finalized_messages(
        &self, 
        message_ids: Vec<IngressMessageId>
    );
}

A component used by Consensus to build and validate a payload.

Required methods

fn get_ingress_payload(
&self,
ingress_pool: &dyn IngressPoolSelect,
past_ingress: &dyn IngressSetQuery,
context: &ValidationContext
) -> IngressPayload

Returns a new ingress payload containing valid Signed Ingress Messages to Consensus.

#Input IngressPool passed as a read-only reference. It's the trait object to read validated messages from the IngressPool. [past_ingress] allows querying if an ingress message exists in past blocks. It is used for deduplication purpose. ValidationContext contains registry_version that allows to validate messages against the correct registry entries, execution_height which is used as parameter for the valid set rule check run to select a valid set of messages.

#Returns IngressPayload which is a collection of valid ingress messages

fn validate_ingress_payload(
&self,
payload: &IngressPayload,
past_ingress: &dyn IngressSetQuery,
context: &ValidationContext
) -> ValidationResult<IngressPayloadValidationError>

Validates an IngressPayload against the past payloads and ValidationContext. The size of the payload is derived from registry.

#Input IngressPayload is the payload to be validated. [past_ingress] allows querying if an ingress message exists in past blocks. It is used for deduplication purpose. ValidationContext Refers to the validation context which contains registry_version, execution_height to be used as a parameter for the valid set rule check run to select a valid set of messages

#Returns ValidationResult::Valid: if the payload is valid ValidationResult::Invalid: if the payload is invalid ValidationResult::Error: a transient error occured during the validation.

fn request_purge_finalized_messages(&self, message_ids: Vec<IngressMessageId>)

Request purge of the given ingress messages from the pool when they have already been included in finalized blocks.

The actual purge is not required to happen immediately.

Loading content...