Module ic_interfaces::ingress_pool

The ingress pool public interface.

Structs

IngressPoolObject
IngressObject is the format stored in the unvalidated/validated sections of the ingress pool. This is basically a wrapper around the SignedIngress message

Enums

ChangeAction
Change actions applicable to the ingress pool.

SelectResult
Indicate whether something should be selected, and whether selection should continue:

Traits

IngressPool
The ingress pool contains all the ingress artifacts received by P2P and produced by the local node.

IngressPoolSelect
A query interface that selects qualifying artifacts from the validated pool.

IngressPoolThrottler
Interface to throttle user ingress messages

MutableIngressPool
Mutable operations on top of IngressPool.

PoolSection
A Pool section is a part of the ingress pool which contains either validated or unvalidated ingress messages Artifacts in the pool are accessible by their id.

Type Definitions

ChangeSet
Change set for processing unvalidated ingress messages

IngressChangeArtifact

UnvalidatedIngressArtifact
Unvalidated ingress artifact

ValidatedIngressArtifact
Validated ingress artifact