Trait ic_interfaces::messaging::XNetPayloadBuilder

pub trait XNetPayloadBuilder: Send + Sync {
    fn get_xnet_payload(
        &self, 
        height: Height, 
        validation_context: &ValidationContext, 
        past_payloads: &[&XNetPayload], 
        byte_limit: NumBytes
    ) -> Result<XNetPayload, XNetPayloadError>;

    fn validate_xnet_payload(
        &self, 
        payload: &XNetPayload, 
        validation_context: &ValidationContext, 
        past_payloads: &[&XNetPayload], 
        byte_limit: NumBytes
    ) -> ValidationResult<XNetPayloadValidationError>;
}

Interface for selecting Streams for inclusion into a Payload.

Required methods

fn get_xnet_payload(
&self,
height: Height,
validation_context: &ValidationContext,
past_payloads: &[&XNetPayload],
byte_limit: NumBytes
) -> Result<XNetPayload, XNetPayloadError>

Produces an XNetPayload of maximum byte size byte_limit that is valid given a ValidationContext (certified height plus registry version) and past_payloads (the XNetPayloads from all blocks above the certified height, in descending block height order).

With the current implementation, if no valid XNetPayload of size 'byte_limit' exists, then the function returns the smallest XNetPayload which is valid. For practical parameters, for the current implementation XNetPayload will no be more than byte_limit+2kB. More details here: https://docs.google.com/document/d/1cC1-U000cts3GHiEf9OUNhp4clKt2XCG08Zxo35xoWw/edit#

Because payload takes time to make, and this function should not block its caller, An implementation can return Pending so that the caller knows to poll again in the future.

fn validate_xnet_payload(
&self,
payload: &XNetPayload,
validation_context: &ValidationContext,
past_payloads: &[&XNetPayload],
byte_limit: NumBytes
) -> ValidationResult<XNetPayloadValidationError>

Checks whether the provided XNetPayload is valid given a ValidationContext (certified height and registry version) and past_payloads (the XNetPayloads from all blocks above the certified height, in descending block height order); and whether its size in bytes does not exceed byte_limit.

Loading content...