Module ic_interfaces::p2p

The P2P public interface.

Traits

IngressEventHandler
This is an event handler that can be used to submit an ingress message to P2P event channels for processing. It encapsulates the given ingress message in a Gossip artifact and sends it to the P2P GossipArtifact channel. It is mainly to be used by the HTTP handler to submit ingress messages.

P2PRunner
P2P exposes channels that are used to hold artifacts sent by the Transport layer or the HTTP handler. These channels also hold any errors and notifications sent by the Transport layer (such as connection/disconnection events). P2PRunner provides the run interface used by the replica to start reading from these channels. The artifacts or notifications received from these channels are sent to Gossip for processing.