Module ic_interfaces::registry

The registry public interface.

Structs

RegistryVersionedRecord
A versioned (Key, Value) pair returned from the registry.

Constants

ZERO_REGISTRY_VERSION
The registry at version 0 is the empty registry.

Traits

LocalStoreCertifiedTimeReader
Whenever the local store is successfully updated, the time contained in the certification is stored on disk. The LocalStoreCertifiedTimeReader provides a method to read this time from disk.

RegistryClient
The RegistryClient provides methods to query the local state of the registry. All methods on this trait return immediately (i.e. there are no side-effects on the critical path).

RegistryDataProvider
A RegistryDataProvider is the data source that backs the RegistryClient, i.e. the registry client uses an instances of this trait to get data from the registry. In production, this trait will be instantiated by an implementation that queries the registry canister on the NNS. For testing and local deployment, this can be instantiated with an implementation that reads from a local file, e.g.

RegistryValue
A Protocol Buffers message.

Functions

empty_zero_registry_record

Type Definitions

RegistryClientResult
Result returns when fetching a value from the registry.

RegistryClientVersionedResult
Result returns when fetching a versioned value from the registry.

RegistryTransportRecord
A RegistryRecord represents a versioned k/v-pair as stored in the registry canister. TODO(IC-102): This needs to be renamed to RegistryRecord after we removed the legacy RegistryRecord.

Derive Macros

RegistryValue