Trait ic_interfaces::registry::RegistryDataProvider

pub trait RegistryDataProvider: Send + Sync {
    fn get_updates_since(
        &self, 
        version: RegistryVersion
    ) -> Result<(Vec<RegistryTransportRecord>, RegistryVersion), RegistryDataProviderError>;
}

A RegistryDataProvider is the data source that backs the RegistryClient, i.e. the registry client uses an instances of this trait to get data from the registry. In production, this trait will be instantiated by an implementation that queries the registry canister on the NNS. For testing and local deployment, this can be instantiated with an implementation that reads from a local file, e.g.

Required methods

fn get_updates_since(
&self,
version: RegistryVersion
) -> Result<(Vec<RegistryTransportRecord>, RegistryVersion), RegistryDataProviderError>

If successful, the call returns a pair, the first element of which is a list of records that were added to the database since version. The second element of the pair is the current registry version.

The record version of the returned records MUST be greater than version and MUST be smaller or equal to the current registry version.

Thus, if the local cache is at version version, after the returned records have been added to the local cache, the local registry is at the current version.

Loading content...