Module ic_interfaces::state_manager

The state manager public interface.

Structs

Labeled
A node state with a height attached to it, indicating that the state was obtained by executing a block with the given height.

Enums

CertificationScope
Indicates the subset of the state that needs to be certified.

CertificationTag

StateManagerError

Constants

CERT_ANY
The mask that accepts all states, not matter what their certification status is.

CERT_CERTIFIED
The mask that only accepts certified states.

CERT_UNCERTIFIED
The mask that only accepts uncertified states.

Traits

StateManager
APIs related to fetching and certifying the state.

StateReader
This component is analogous to the StateManager except that it is used to access State which cannot be checkpointed or snapshotted.

Type Definitions

CertificationMask
A bit mask for accepting certified/uncertified states.

StateManagerResult