Trait ic_interfaces::state_manager::StateManager

pub trait StateManager: StateReader {
    fn list_state_hashes_to_certify(
        &self
    ) -> Vec<(Height, CryptoHashOfPartialState)>;

    fn deliver_state_certification(&self, certification: Certification);

    fn get_state_hash_at(
        &self, 
        height: Height
    ) -> StateManagerResult<Option<CryptoHashOfState>>;

    fn fetch_state(&self, height: Height, root_hash: CryptoHashOfState);

    fn list_state_heights(&self, cert_mask: CertificationMask) -> Vec<Height>;

    fn remove_states_below(&self, height: Height);

    fn commit_and_certify(
        &self, 
        state: Self::State, 
        height: Height, 
        scope: CertificationScope
    );

    fn take_tip(&self) -> (Height, Self::State);

    fn take_tip_at(&self, height: Height) -> StateManagerResult<Self::State>;

    fn report_diverged_state(&self, height: Height);
}

APIs related to fetching and certifying the state.

Required methods

fn list_state_hashes_to_certify(
&self
) -> Vec<(Height, CryptoHashOfPartialState)>

Returns a snapshot of the list of state hashes that need to be certified ("the list" below).

The actual list is maintained by the StateManager. The following operations can modify the list:

A call to commit_and_certify starts a potentially asynchronous computation that adds an entry to the list. This implies that the list will contain every height at some point, unless the state is removed before its hash is computed.

A call to deliver_state_certification with a certification of some height removes the corresponding entry from the list.

A call to one of the remove_state* methods removes the corresponding entry from the list.

Since the hash computation can be asynchronous, the order in which heights appear in the list can differ from the order in which states are committed. E.g., the following outcome is possible:

sm.commit_and_certify(state_1, h_1, scope_1)
sm.commit_and_certify(state_1, h_2, scope_1)
sm.list_state_hashes_to_certify() = [(h_2, H_2)]
sm.list_state_hashes_to_certify() = [(h_1, H_1), (h_2, H_2)]

Properties

The entries in the list are unique and sorted by height.

let l = state_manager.list_state_hashes_to_certify()
∀ i, j ∈ indices(l): i < j ⇒ l[i].height < l[j].height

The hash associated with a height is guaranteed not to change.

let l_1 = state_manager.list_state_hashes_to_certify()
...
let l_2 = state_manager.list_state_hashes_to_certify()
∀ (h_1, H_1) ∈ l_1, (h_2, H_2) ∈ l_2: h_1 = h_2 ⇒ H_1 = H_2

fn deliver_state_certification(&self, certification: Certification)

Delivers a certification corresponding to some state hash / height pair.

Does nothing if another certification for the same state has already been delivered.

Post-conditions

(height, ·) ∉ state_manager.list_state_hashes_to_certify()

Panics

Panics if certification.content.hash is not equal to the hash computed from the state at height certification.content.height.

fn get_state_hash_at(
&self,
height: Height
) -> StateManagerResult<Option<CryptoHashOfState>>

Returns the hash of the state at the specified height.

The state hash can be computed asynchronously. If the state itself is available, but the hash is not computed yet, this method returns Ok(None).

Errors

If the state at height was already removed, the StateRemoved error is returned.

If the state at height is not committed yet, the StateNotCommittedYet error is returned.

fn fetch_state(&self, height: Height, root_hash: CryptoHashOfState)

Initiates asynchronous procedure of state synchronization with the target state specified by its height and root hash.

Does nothing if self.latest_state_height() >= height.

Note that there is no explicit notification (or callback) indicating that the fetch is complete. The caller is supposed to poll self.latest_state_height() periodically to learn that the state became available.

Panics

Panics if the state is already known and its root_hash differs, i.e. self.get_state_hash_at(height) = Ok(Some(h)) ∧ h ≠ root_hash

fn list_state_heights(&self, cert_mask: CertificationMask) -> Vec<Height>

Returns the list of heights corresponding to accessible states matching the mask. E.g. list_state_heights(CERT_ANY) will return all accessible states.

Note that the initial state at height 0 is considered uncertified from the State Manager point of view. This is because the protocol requires each replica to individually obtain the initial state using some out-of-band mechanism (i.e., not state sync). Also note that the authenticity of this initial state will be verified by some protocol external to this component.

The list of heights is guaranteed to be

Non-empty if cert_mask = CERT_ANY as it will contain at least height 0 even if no states were committed yet.

Sorted in ascending order.

fn remove_states_below(&self, height: Height)

Notify this state manager that states with heights strictly less than the specified height can be removed.

Note that:

The initial state (height = 0) cannot be removed.

Some states matching the removal criteria might be kept alive. For example, the last fully persisted state might be preserved to optimize future operations.

fn commit_and_certify(
&self,
state: Self::State,
height: Height,
scope: CertificationScope
)

Commits the state at given height, limits the certification to scope. The state must be the mutable state obtained via a call to take_tip.

Does nothing if height ≤ state_manager.latest_state_height().

Panics

Panics if the state at height has already been committed before but has a different hash.

fn take_tip(&self) -> (Height, Self::State)

Returns the version of the state that can be modified in-place and the height of that state.

This function transfers the ownership of the mutable state from StateManager to the caller.

Panics

This function panics if StateManager doesn't own the mutable state. That means that take_tip cannot be called twice in a row, every invocation of take_tip must be balanced by a call to commit_and_certify.

fn take_tip_at(&self, height: Height) -> StateManagerResult<Self::State>

Returns the version of the state that can be modified in-place (TIP) if the requested height matches height(TIP).

Errors

If the height < height(TIP), returns StateRemoved.

If the height(TIP) < height, returns StateNotCommittedYet.

Panics

This function panics if StateManager doesn't own the TIP. See take_tip for more details.

fn report_diverged_state(&self, height: Height)

Reports that the given height diverged, i.e., consensus agreed on the hash of state at height, and this was not the state produced by this state manager.

This function triggers non-determinism recovery procedure, which involves pruning all the states that might be diverged as well.

Panics

This function always panics because there is no point in continuing the normal operation. We rely on node manager restarting the replica, which in turn will initiate the normal recovery procedure.

Loading content...