Trait ic_interfaces::state_manager::StateReader

pub trait StateReader: Send + Sync {
    type State;
    fn get_state_at(
        &self, 
        height: Height
    ) -> StateManagerResult<Labeled<Arc<Self::State>>>;

    fn get_latest_state(&self) -> Labeled<Arc<Self::State>>;

    fn latest_state_height(&self) -> Height;

    fn latest_certified_height(&self) -> Height;

    fn read_certified_state(
        &self, 
        paths: &LabeledTree<()>
    ) -> Option<(Arc<Self::State>, MixedHashTree, Certification)>;
}

This component is analogous to the StateManager except that it is used to access State which cannot be checkpointed or snapshotted.

Associated Types

type State

Type of state managed by StateReader.

Should typically be ic_replicated_state::ReplicatedState.

Loading content...
Required methods

fn get_state_at(
&self,
height: Height
) -> StateManagerResult<Labeled<Arc<Self::State>>>

Returns a shared object of the state at the specified height. If the state is not available locally, the method blocks until the state is fetched.

Errors

If the state at height was already removed, the StateRemoved error is returned.

If the state at height is not committed yet, the StateNotCommittedYet error is returned.

fn get_latest_state(&self) -> Labeled<Arc<Self::State>>

Returns a shared object of the state at the latest committed block height. If nothing was committed so far, returns an empty valid state.

fn latest_state_height(&self) -> Height

Returns the height of the latest state available.

fn latest_certified_height(&self) -> Height

Returns the height of the latest certified state available.

fn read_certified_state(
&self,
paths: &LabeledTree<()>
) -> Option<(Arc<Self::State>, MixedHashTree, Certification)>

Reads part of the certified state tree specified by the shape of paths. Path can reference either a leaf or a subtree. E.g., if the tree looks like this:

 * -+- time - 1000
    |
    `- request_status -+- 1 - status - processing
                       |
                       `- 2 - status - processing

then the allowed paths are:

/time

/request_status

/request_status/1

/request_status/2

/request_status/1/status

/request_status/2/status

Use Leaf(()) to indicate "subtree rooted at this path", e.g., SubTree { "request_status" => SubTree { "1" => Leaf(()) } } is a subtree that requests all the nodes below path /request_status/1.

Returns None if there is no certified state available yet.

Loading content...