Trait ic_interfaces::transport::Transport

pub trait Transport: Send + Sync {
    fn register_client(
        &self, 
        client_type: TransportClientType, 
        async_event_handler: Arc<dyn AsyncTransportEventHandler>
    ) -> Result<(), TransportErrorCode>;

    fn start_connections(
        &self, 
        client_type: TransportClientType, 
        peer: &NodeId, 
        node_record: &NodeRecord, 
        registry_version: RegistryVersion
    ) -> Result<(), TransportErrorCode>;

    fn stop_connections(
        &self, 
        client_type: TransportClientType, 
        peer_id: &NodeId, 
        registry_version: RegistryVersion
    ) -> Result<(), TransportErrorCode>;

    fn send(
        &self, 
        client_type: TransportClientType, 
        peer_id: &NodeId, 
        flow_tag: FlowTag, 
        message: TransportPayload
    ) -> Result<(), TransportErrorCode>;

    fn clear_send_queues(
        &self, 
        client_type: TransportClientType, 
        peer_id: &NodeId
    );

    fn clear_send_queue(
        &self, 
        client_type: TransportClientType, 
        peer_id: &NodeId, 
        flow_tag: FlowTag
    );
}

Transport layer APIs.

Required methods

fn register_client(
&self,
client_type: TransportClientType,
async_event_handler: Arc<dyn AsyncTransportEventHandler>
) -> Result<(), TransportErrorCode>

Register the transport client of the specified type. No more than one module can register against a particular client type. This returns a handle to the client's context, which should be supplied in all the future interactions with the transport layer.

fn start_connections(
&self,
client_type: TransportClientType,
peer: &NodeId,
node_record: &NodeRecord,
registry_version: RegistryVersion
) -> Result<(), TransportErrorCode>

Mark the peer as valid neighbor, and set up the transport layer to exchange messages with the peer. This call would create the necessary wiring in the transport layer for the peer:

Set up the Tx/Rx queueing, based on TransportQueueConfig.

If the peer is the server, initiate connection requests to the peer server ports.

If the peer is the client, set up the connection state to accept connection requests from the peer.

fn stop_connections(
&self,
client_type: TransportClientType,
peer_id: &NodeId,
registry_version: RegistryVersion
) -> Result<(), TransportErrorCode>

Remove the peer from the set of valid neighbors, and tear down the queues and connections for the peer. Any messages in the Tx and Rx queues for the peer will be discarded.

fn send(
&self,
client_type: TransportClientType,
peer_id: &NodeId,
flow_tag: FlowTag,
message: TransportPayload
) -> Result<(), TransportErrorCode>

Send the message to the specified peer. The message will be enqueued into the appropriate TxQ based on the TransportQueueConfig.

fn clear_send_queues(&self, client_type: TransportClientType, peer_id: &NodeId)

Clear any unsent messages in all the send queues for the peer.

fn clear_send_queue(
&self,
client_type: TransportClientType,
peer_id: &NodeId,
flow_tag: FlowTag
)

Clear any unsent messages in the send queue for the given peer and flow tag.

Loading content...