Trait ic_interfaces::transport::TransportEventHandler

pub trait TransportEventHandler: Send + Sync {
    fn on_message(
        &self, 
        flow: FlowId, 
        message: TransportPayload
    ) -> Option<TransportPayload>;

    fn on_error(&self, flow: FlowId, error: TransportErrorCode);

    fn on_state_change(&self, state_change: TransportStateChange);
}

API for handling transport events.

Required methods

fn on_message(
&self,
flow: FlowId,
message: TransportPayload
) -> Option<TransportPayload>

Invoked by the transport layer when a message is received from the network. This is implemented by the transport clients to receive/process the messages. Returns the message back if it was not accepted.

fn on_error(&self, flow: FlowId, error: TransportErrorCode)

fn on_state_change(&self, state_change: TransportStateChange)

Invoked by the transport layer to notify of any changes in the state.

Loading content...