Enum ic_interfaces::validation::ValidationError

pub enum ValidationError<P, T> {
    Permanent(P),
    Transient(T),
}

Validation errors can be either permanent (invalid) or transient.

Variants

Permanent(P)

Transient(T)

Implementations

impl<P, T> ValidationError<P, T>

pub fn map<Q, S, F: Fn(P) -> Q, G: Fn(T) -> S>(
self,
f: F,
g: G
) -> ValidationError<Q, S>

The inner types of ValidationError can be mapped to another types by applying two map functions for error and transient errors respectively.